snapshot: add swc-loader in separate webpack config
The massive surge forward in progress (as reported by WebpackBar) is
swc-loader doing its highly-optimized thing.  Unfortunately, there's
just so much other stuff going on around it that a build still takes
around 30 seconds.

Because swc-loader doesn't support babel plugins (such as the
`babel-plugin-import` we rely on for antd imports), this benchmark
stacks on top of the esbuild prototype.  It shows a valid comparison
between the two options, but probably has a slight advantage compared to
our vanilla builds in `master` that use `babel-plugin-import` expand a
single top-level import from antd into a component + style import.